devrequirements.txt
-c constraints.txt
pylint
pylint-django
yapf